+++ source/extensions/isaacsim.replicator.examples/python/tests/test_sdg_ur10_palletizing.py
@@ -16,6 +16,7 @@
 import os
 import unittest
 
+import carb.settings
 import omni.kit
 import omni.usd
 from isaacsim.test.utils.file_validation import validate_folder_contents
@@ -26,6 +27,7 @@ class TestSDGUR10Palletizing(omni.kit.test.AsyncTestCase):
         await omni.kit.app.get_app().next_update_async()
         await omni.usd.get_context().new_stage_async()
         await omni.kit.app.get_app().next_update_async()
+        self.original_dlss_exec_mode = carb.settings.get_settings().get("rtx/post/dlss/execMode")
 
     async def tearDown(self):
         omni.usd.get_context().close_stage()
@@ -33,6 +35,7 @@ class TestSDGUR10Palletizing(omni.kit.test.AsyncTestCase):
         # In some cases the test will end before the asset is loaded, in this case wait for assets to load
         while omni.usd.get_context().get_stage_loading_status()[2] > 0:
             await omni.kit.app.get_app().next_update_async()
+        carb.settings.get_settings().set("rtx/post/dlss/execMode", self.original_dlss_exec_mode)
 
     @unittest.skipIf(os.getenv("ETM_ACTIVE"), "Skipped in ETM.")
     async def test_sdg_ur10_palletizing(self):
@@ -116,6 +119,9 @@ class TestSDGUR10Palletizing(omni.kit.test.AsyncTestCase):
 
                 # Disable capture on play for replicator, data capture will be triggered manually
                 rep.orchestrator.set_capture_on_play(False)
+
+                # Set DLSS to Quality mode (2) for best SDG results , options: 0 (Performance), 1 (Balanced), 2 (Quality), 3 (Auto)
+                carb.settings.get_settings().set("rtx/post/dlss/execMode", 2)
 
                 # Clear any previously generated SDG graphs
                 if self._stage.GetPrimAtPath("/Replicator"):