+++ source/extensions/isaacsim.replicator.examples/python/tests/test_sdg_useful_snippets_timeline_based.py
@@ -13,8 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
-
+import carb.settings
 import omni.kit
 import omni.usd
 from isaacsim.test.utils.file_validation import validate_folder_contents
@@ -25,6 +24,7 @@ class TestSDGUsefulSnippets(omni.kit.test.AsyncTestCase):
         await omni.kit.app.get_app().next_update_async()
         omni.usd.get_context().new_stage()
         await omni.kit.app.get_app().next_update_async()
+        self.original_dlss_exec_mode = carb.settings.get_settings().get("rtx/post/dlss/execMode")
 
     async def tearDown(self):
         omni.usd.get_context().close_stage()
@@ -32,6 +32,7 @@ class TestSDGUsefulSnippets(omni.kit.test.AsyncTestCase):
         # In some cases the test will end before the asset is loaded, in this case wait for assets to load
         while omni.usd.get_context().get_stage_loading_status()[2] > 0:
             await omni.kit.app.get_app().next_update_async()
+        carb.settings.get_settings().set("rtx/post/dlss/execMode", self.original_dlss_exec_mode)
 
     async def test_sdg_snippet_custom_fps_writer_annotator(self):
         import asyncio
@@ -57,6 +58,9 @@ class TestSDGUsefulSnippets(omni.kit.test.AsyncTestCase):
 
             # Make sure fixed time stepping is set (the timeline will be advanced with the same delta time)
             carb.settings.get_settings().set("/app/player/useFixedTimeStepping", True)
+
+            # Set DLSS to Quality mode (2) for best SDG results , options: 0 (Performance), 1 (Balanced), 2 (Quality), 3 (Auto)
+            carb.settings.get_settings().set("rtx/post/dlss/execMode", 2)
 
             # Set the timeline parameters
             timeline = omni.timeline.get_timeline_interface()
@@ -106,7 +110,7 @@ class TestSDGUsefulSnippets(omni.kit.test.AsyncTestCase):
                     rp.hydra_texture.set_updates_enabled(True)
 
                     # Step needs to be called after scheduling the write
-                    await rep.orchestrator.step_async(delta_time=0.0, pause_timeline=False)
+                    await rep.orchestrator.step_async(delta_time=0.0, pause_timeline=False, rt_subframes=16)
 
                     # After step, the annotator data is available and in sync with the stage
                     annot_data = annot_depth.get_data()