+++ source/standalone_examples/replicator/augmentation/writer_augmentation.py
@@ -42,6 +42,9 @@ ENV_URL = "/Isaac/Environments/Grid/default_environment.usd"
 # Enable scripts
 carb.settings.get_settings().set_bool("/app/omni.graph.scriptnode/opt_in", True)
 
+# Set DLSS to Quality mode (2) for best SDG results , options: 0 (Performance), 1 (Balanced), 2 (Quality), 3 (Auto)
+carb.settings.get_settings().set("rtx/post/dlss/execMode", 2)
+
 
 # Gaussian noise augmentation on rgba data in numpy (CPU) and warp (GPU)
 def gaussian_noise_rgb_np(data_in, sigma: float, seed: int):