+++ source/standalone_examples/replicator/infinigen/infinigen_sdg.py
@@ -202,6 +202,9 @@ def run_sdg(config):
     # Disable UJITSO cooking ([Warning] [omni.ujitso] UJITSO : Build storage validation failed)
     carb.settings.get_settings().set("/physics/cooking/ujitsoCollisionCooking", False)
 
+    # Set DLSS to Quality mode (2) for best SDG results , options: 0 (Performance), 1 (Balanced), 2 (Quality), 3 (Auto)
+    carb.settings.get_settings().set("rtx/post/dlss/execMode", 2)
+
     # Debug mode (hide ceiling, move viewport camera to the top-down view)
     debug_mode = config.get("debug_mode", False)